Refactor test_pages.py
